create a serverless trace flusher, that just immediately sends traces for now
